refactor: change file directories
